requirements.txt
python-telegram-bot
asyncio
python-dotenv
requests
